scripts/requirements.txt
google-cloud-aiplatform>=1.38.0,<2.0.0
tiktoken>=0.5.0

